refactor: move getManualIssueRequests
